pyproject.toml
[project]
name = "convert-image-format"
version = "0.1.0"
description = "A versatile image conversion tool supporting various formats, including HEIC/HEIF and camera RAW files."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.0",
    "pillow>=11.2.1",
    "pillow-heif>=0.22.0",
    "rawpy>=0.25.0",
]
